debug

+++ scripts/stop.sh
@@ -2,5 +2,9 @@
 
 # stop docker compose 
 docker-compose down
-# delete all containers
-docker rm $(docker ps -a -q)
+# before removing all containers, check if there are any containers running, if so, stop them and remove them
+if [ "$(docker ps -q)" ]; then
+    docker rm $(docker ps -a -q)
+fi
+
+
